+++ src/main/php/io/streams/compress/Algorithm.class.php
@@ -16,9 +16,12 @@ interface Algorithm {
   /** Returns the algorithm's common file extension, including a leading "." */
   public function extension(): string;
 
+  /** Maps fastest, default and strongest levels */
+  public function level(int $select): int;
+
   /** Opens an input stream for reading */
   public function open(InputStream $in): InputStream;
 
   /** Opens an output stream for writing */
-  public function create(OutputStream $out, int $method= Compression::DEFAULT): OutputStream;
+  public function create(OutputStream $out, int $level= Compression::DEFAULT): OutputStream;
 }