+++ src/main/php/io/streams/compress/Brotli.class.php
@@ -16,15 +16,19 @@ class Brotli implements Algorithm {
   /** Returns the algorithm's common file extension, including a leading "." */
   public function extension(): string { return '.br'; }
 
+  /** Maps fastest, default and strongest levels */
+  public function level(int $select): int {
+    static $levels= [Compression::FASTEST => 1, Compression::DEFAULT => 11, Compression::STRONGEST => 11];
+    return $levels[$select] ?? $select;
+  }
+
   /** Opens an input stream for reading */
   public function open(InputStream $in): InputStream {
     return new BrotliInputStream($in);
   }
 
   /** Opens an output stream for writing */
-  public function create(OutputStream $out, int $method= Compression::DEFAULT): OutputStream {
-    static $levels= [Compression::FASTEST => 1, Compression::DEFAULT => 11, Compression::STRONGEST => 11];
-
-    return new BrotliOutputStream($out, $levels[$method]);
+  public function create(OutputStream $out, int $level= Compression::DEFAULT): OutputStream {
+    return new BrotliOutputStream($out, $this->level($level));
   }
 }